add Product, Import Product, Sell Product, Test JUnit on set sale price

+++ src/main/java/com/piseth/example/spring/phone_shop/repository/BrandRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface BrandRepository extends JpaRepository<Brand, Integer>, JpaSpecificationExecutor<Brand> {
+public interface BrandRepository extends JpaRepository<Brand, Long>, JpaSpecificationExecutor<Brand> {
     List<Brand> findByNameLike(String name);
 
     // Homework for do junit test with h2 database
